add address data to country list
There is address information in each of the locales, but this will
allow us to generate a foreign address without changing the locale.

For example, if the current locale is :en and I want a french address,
I must first change to the :fr locale before calling the
Faker::Address methods, then remember to change it back. This is
suboptimal.

Note that because I'm porting this change from a repository that is
several years old, there is data in newer localizations that isn't in
this list. Someone should get them in sync, but I haven't the time.